+++ config/config.exs
@@ -64,8 +64,7 @@ config :td_audit, :phoenix_swagger,
 config :td_audit, permission_resolver: TdPerms.Permissions
 
 config :td_audit, user_cache: TdPerms.UserCache
-
-config :td_audit, business_concept_cache: TdPerms.BusinessConceptCache
+config :td_audit, bc_cache: TdPerms.BusinessConceptCache
 
 config :td_audit, concepts_path: "/concepts"