+++ lib/td_audit/search/search_event_processor.ex
@@ -2,20 +2,29 @@ defmodule TdAudit.SearchEventProcessor do
   require Logger
   @moduledoc false
   alias TdAudit.BusinessConcept.Search
-  alias TdPerms.BusinessConceptCache
 
-  def process_event(%{"event" => "create_rule", "payload" => payload}) do
-    resource_id = Map.get(payload, "business_concept_id")
-    BusinessConceptCache.increment(resource_id, "rule_count")
+  @bc_cache Application.get_env(:td_audit, :bc_cache)
+
+  def process_event(%{
+    "event" => "create_rule",
+    "payload" => %{
+      "business_concept_id" => resource_id
+    }
+  }) do
+    @bc_cache.increment(resource_id, "rule_count")
 
     Search.update_business_concept_by_script(
       %{business_concept_id: resource_id},
       retrieve_script_map("rule_count").increment_int_field
     )
   end
-  def process_event(%{"event" => "delete_rule", "payload" => payload}) do
-    resource_id = Map.get(payload, "business_concept_id")
-    BusinessConceptCache.decrement(resource_id, "rule_count")
+  def process_event(%{
+    "event" => "delete_rule",
+    "payload" => %{
+      "business_concept_id" => resource_id
+    }
+  }) do
+    @bc_cache.decrement(resource_id, "rule_count")
 
     Search.update_business_concept_by_script(
       %{business_concept_id: resource_id},
@@ -23,10 +32,11 @@ defmodule TdAudit.SearchEventProcessor do
     )
   end
   def process_event(%{
-        "event" => "add_relation",
-        "resource_id" => resource_id
-      }) do
-    BusinessConceptCache.increment(resource_id, "link_count")
+    "event" => "add_relation",
+    "resource_id" => resource_id,
+    "payload" => %{"target_type" => "data_field"}
+  }) do
+    @bc_cache.increment(resource_id, "link_count")
 
     Search.update_business_concept_by_script(
       %{business_concept_id: resource_id},
@@ -34,10 +44,11 @@ defmodule TdAudit.SearchEventProcessor do
     )
   end
   def process_event(%{
-        "event" => "delete_relation",
-        "resource_id" => resource_id
-      }) do
-    BusinessConceptCache.decrement(resource_id, "link_count")
+    "event" => "delete_relation",
+    "resource_id" => resource_id,
+    "payload" => %{"target_type" => "data_field"}
+  }) do
+    @bc_cache.decrement(resource_id, "link_count")
 
     Search.update_business_concept_by_script(
       %{business_concept_id: resource_id},