+++ lib/td_audit_web/controllers/fallback_controller.ex
@@ -9,12 +9,14 @@ defmodule TdAuditWeb.FallbackController do
   def call(conn, {:error, %Ecto.Changeset{} = changeset}) do
     conn
     |> put_status(:unprocessable_entity)
-    |> render(TdAuditWeb.ChangesetView, "error.json", changeset: changeset)
+    |> put_view(TdAuditWeb.ChangesetView)
+    |> render("error.json", changeset: changeset)
   end
 
   def call(conn, {:error, :not_found}) do
     conn
     |> put_status(:not_found)
-    |> render(TdAuditWeb.ErrorView, :"404")
+    |> put_view(TdAuditWeb.ErrorView)
+    |> render("404.json")
   end
 end